requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
rembg==2.0.67
Pillow==10.0.0
torch==2.1.0
ultralytics==8.0.117
